# app/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zhixue Mate Agent API",
    "version": "api-contract-v0.3",
    "description": "知学 Mate 前后端唯一接口契约。字段变更必须先升契约版本并通知前端。 声明层已于 2026-09-22 按实跑结果补全：补齐 5 个已实现端点、所有 operation 的 405、register 的 409、exercises 的 4 个查询参数与 400、knowledge-bases 的 400、可空字段的 nullable、版本头参数，并移除死枚举值 LLM_FALLBACK。实现未改动。 KnowledgeBaseDetailResponse.documents 已于 2026-09-22 补声明（后端返回、前端在读，此前契约未声明）。 LegacyChatResponse.card.summary 已于 2026-09-22 补声明（前端在读、后端已补返回，此前两边都缺）。"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发"
    },
    {
      "url": "http://{host}:5000",
      "description": "局域网联调",
      "variables": {
        "host": {
          "default": "192.168.1.100"
        }
      }
    }
  ],
  "security": [],
  "paths": {
...
    "/api/v1/plans/current": {
      "get": {
        "summary": "读取当前学习计划",
        "operationId": "getCurrentPlan",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdQuery"
          },
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "当前计划",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{setId}/submit": {
      "post": {
        "summary": "提交练习并完成评估",
        "operationId": "submitExercises",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExerciseSetId"
          },
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseSubmitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "评估结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSubmitResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{setId}": {
      "get": {
        "summary": "获取练习题集",
        "operationId": "getExerciseSet",
        "parameters": [
          {
            "$ref": "#/components/parameters/ExerciseSetId"
          },
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          },
          {
            "name": "generate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "题库未覆盖该知识点时是否允许服务端**现场生成**题目。传 1/true/yes 允许；缺省不生成（行为与之前完全一致）。生成题的 source 为 llm，**判分仍是确定性算式**（按存下的 answerKey 逐题比对）。"
          },
          {
            "name": "knowledgePointId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "按知识点过滤。与 difficulty/excludeExerciseId 任一存在时走筛选路径。"
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "按难度过滤（如 easy / medium / hard）。"
          },
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "返回题目数量上限，必须为正整数；非整数返回 400。缺省取 3。"
          },
          {
            "name": "excludeExerciseId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "要排除的题目 ID。可重复传递以排除多题。",
            "style": "form",
            "explode": true
          }
        ],
        "responses": {
          "200": {
            "description": "练习题集",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseSetResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "count 非正整数等参数错误（errorCode=BAD_REQUEST）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "可选鉴权。/api/v1/** 不带 token 时按演示/游客身份（demo-user）处理，以保证演示主链不受影响。"
      }
    },
    "parameters": {
...
      "UserIdQuery": {
        "name": "userId",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "default": "demo-user"
        }
      },
      "ExerciseSetId": {
        "name": "setId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "ContractVersionHeader": {
        "name": "X-API-Contract-Version",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "接口契约版本闸门。**可选**：不带头时后端放行；带上但值不等于当前版本时返回 409 CONTRACT_VERSION_MISMATCH。当前版本见 info.version。"
      }
    },
    "schemas": {
...
      "PlanTask": {
        "type": "object",
        "required": [
          "taskId",
          "knowledgePointId",
          "knowledgePointName",
          "durationMinutes",
          "status"
        ],
        "properties": {
          "taskId": {
            "type": "string"
          },
          "knowledgePointId": {
            "type": "string"
          },
          "knowledgePointName": {
            "type": "string"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "completed"
            ]
          },
          "priority": {
            "type": "string"
          }
        }
      },
      "LearningPlan": {
        "type": "object",
        "required": [
          "planId",
          "version",
          "tasks",
          "generatedFromProfileVersion"
        ],
        "properties": {
          "planId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanTask"
            }
          },
          "factors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriorityFactor"
            }
          },
          "reason": {
            "type": "string"
          },
          "generatedFromProfileVersion": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "PlanResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LearningPlan"
          }
        ]
      },
      "Exercise": {
        "type": "object",
        "required": [
          "exerciseId",
          "knowledgePointId",
          "knowledgePointName",
          "stem",
          "options"
        ],
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "knowledgePointId": {
            "type": "string"
          },
          "knowledgePointName": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "stem": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string"
          }
        }
      },
      "ExerciseSetResponse": {
        "type": "object",
        "required": [
          "setId",
          "exercises"
        ],
        "properties": {
          "setId": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exercise"
            },
            "description": "固定 Demo 返回 3 道题；无可用题目时允许为空数组。"
          }
        }
      },
      "ExerciseSubmitRequest": {
        "type": "object",
        "required": [
          "idempotencyKey",
          "answers"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "default": "demo-user"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "idempotencyKey": {
            "type": "string",
            "description": "客户端为本次提交生成的稳定唯一键；重复使用同一键不会产生重复写入。"
          },
          "answers": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "exerciseId",
                "answer"
              ],
              "properties": {
                "exerciseId": {
                  "type": "string"
                },
                "answer": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "AssessmentResult": {
        "type": "object",
        "required": [
          "score",
          "perKnowledgeAccuracy",
          "errorTypes",
          "oldMastery",
          "suggestedNewMastery"
        ],
        "properties": {
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "perKnowledgeAccuracy": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "errorTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "oldMastery": {
            "type": "number"
          },
          "suggestedNewMastery": {
            "type": "number"
          }
        }
      },
      "ExerciseSubmitResponse": {
        "type": "object",
        "required": [
          "userId",
          "submissionId",
          "assessment",
          "masteryUpdate",
          "needReplan",
          "traceId",
          "evidenceIds",
          "replanDecision"
        ],
        "properties": {
          "userId": {
            "type": "string"
          },
          "submissionId": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "evidenceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replanDecision": {
            "$ref": "#/components/schemas/ReplanDecision"
          },
          "assessment": {
            "$ref": "#/components/schemas/AssessmentResult"
          },
          "masteryUpdate": {
            "type": "object",
            "required": [
              "knowledgePointId",
              "oldScore",
              "newScore"
            ],
            "properties": {
              "knowledgePointId": {
                "type": "string"
              },
              "oldScore": {
                "type": "number"
              },
              "newScore": {
                "type": "number"
              }
            }
          },
          "needReplan": {
            "type": "boolean"
          },
          "plan": {
            "$ref": "#/components/schemas/LearningPlan"
          },
          "planDiff": {
            "$ref": "#/components/schemas/PlanDiffResponse"
          }
        }
      },
      "ReplanDecision": {
        "type": "object",
        "required": [
          "needReplan",
          "reasons"
        ],
        "properties": {
          "needReplan": {
            "type": "boolean"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PriorityFactor": {
        "type": "object",
        "required": [
          "name",
          "value",
          "weight",
          "contribution"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "归一化为 0~1。"
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "contribution": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
...
      "PlanDiffResponse": {
        "type": "object",
        "required": [
          "oldVersion",
          "newVersion",
          "changedTasks",
          "adjustmentReason",
          "triggerEvidence"
        ],
        "properties": {
          "planId": {
            "type": "string"
          },
          "oldVersion": {
            "type": "integer"
          },
          "newVersion": {
            "type": "integer"
          },
          "changedTasks": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "adjustmentReason": {
            "type": "string"
          },
          "triggerEvidence": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "properties": {
          "errorCode": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "CONTRACT_VERSION_MISMATCH",
              "INTERNAL_ERROR",
              "RATE_LIMITED",
              "DAILY_LIMIT_REACHED",
              "PHONE_LOCKED",
              "INVALID_CODE",
              "CODE_EXPIRED",
              "CODE_NOT_FOUND",
              "SMS_NOT_CONFIGURED",
              "SMS_PROVIDER_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "幂等键冲突或状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}